+++ new-client/src/plugins/Anchor/Anchor.js
@@ -37,6 +37,9 @@ class Anchor extends React.PureComponent {
           globalObserver={this.props.app.globalObserver}
           model={this.props.app.anchorModel}
           options={this.props.options}
+          enableAppStateInHash={
+            this.props.app.config.mapConfig.map.enableAppStateInHash
+          }
         />
       </BaseWindowPlugin>
     );